added confirm delete box using sweetalert and ajax

+++ admin/p_code.php
@@ -32,17 +32,9 @@ if ($_SERVER["REQUEST_METHOD"] == "POST" && isset($_POST['product_edit']) ) {
     updateProduct($conn,$id,$edit_name,$edit_description,$edit_price,$edit_image);
 }
 // DELETE PRODUCT
-// if ( isset($_REQUEST['delete']) ) {
-   $id = $_GET['id'];
-   
-   $delete_query = "DELETE FROM product WHERE id=$id LIMIT 1";
-   $delete_result =mysqli_query($conn,$delete_query);
-   if($delete_result){
-    $_SESSION['status'] = "Successfully delete product";
-        echo "<script>window.location='product.php'</script>";
-   }else{
-    $_SESSION['status'] = "Can't delete product!";
-    echo "<script>window.location='product.php'</script>";
-   }
-// }
+if (isset($_REQUEST['delete_btn_set'])) {
+    $id = $_POST['delete_id'];
+    $delete_query = "DELETE FROM product WHERE id=$id LIMIT 1";
+    $delete_result = mysqli_query($conn, $delete_query);
+}
     ?>
